Reorganize tests to match code structure
There's now a separate `test_utils` package for things that live in
`utils`, and a new `test_helpers.py` file for testing the test helpers.
